+++ changelog.md
@@ -1,3 +1,7 @@
+# Test coverage (0.0.2)
+#### Tests
+* Unit test for GetTransactionController class is added.
+
 # Initial version (0.0.1)
 #### Features
 * Project structure is inited is MVC manner (_/src/*_).